Improve Satus types
Some of the fields of status are optional, so mark them as such so that
serde does not fail when they are missing.

Most fields are numbers so convert them to u64.

I did some prelimeninary poking around the hostapd 2.10 source code to
see what is optional. It is possible older versions may not return all
fields.